refactor: allowance mutation hooks
